refactor(assets-controller): merge wallet-balance selectors into balance.ts
Combine the wallet-/group-level balance calculators and their types into
the existing balance.ts module (and tests into balance.test.ts) per PR
review, removing the separate wallet-balance.ts file.
